requirements.txt
asgiref==3.7.2
astroid==3.3.5
bandit==1.8.0
coverage==7.6.8
diff-match-patch==20241021
dill==0.3.9
Django==4.2.3
django-ckeditor==6.6.1
django-etc==1.4.0
django-hitcount==1.3.5
django-import-export==4.3.3
django-js-asset==2.1.0
gunicorn==20.1.0
iniconfig==2.0.0
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==24.2
pbr==6.1.0
Pillow==10.0.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
Pygments==2.18.0
pylint==3.3.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
pytils==0.4.1
PyYAML==6.0.2
redis==4.5.5
rich==13.9.4
sqlparse==0.4.4
stevedore==5.4.0
tablib==3.7.0
tomlkit==0.13.2
